Split interfaces into separate classes
