+++ src/FlatSharp/TypeModel/ITableTypeModel.cs
@@ -1,0 +1,29 @@
+﻿/*
+ * Copyright 2020 James Courtney
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+namespace FlatSharp.TypeModel
+{
+    /// <summary>
+    /// A table type model.
+    /// </summary>
+    public interface ITableTypeModel : ITypeModel
+    {
+        /// <summary>
+        /// When set, indicates the table member that acts as the key for sorting vectors containing this table.
+        /// </summary>
+        TableMemberModel KeyMember { get; }
+    }
+}
